+++ app/src/main/java/com/hotmail/leon/zimmermann/homeassistant/models/tables/category/Category.kt
@@ -12,5 +12,5 @@ enum class Category {
     DRINKS,
     BABY;
 
-    val id: Int = ordinal
+    val id: Long = ordinal.toLong()
 }
