+++ app/src/main/java/com/hotmail/leon/zimmermann/homeassistant/models/tables/measure/Measure.kt
@@ -14,7 +14,7 @@ enum class Measure(
     DECILITER("deciliter", "dl", MeasureType.VOLUME, 0.1),
     LITER("liter", "l", MeasureType.VOLUME, 1.0);
 
-    val id: Int = ordinal
+    val id: Long = ordinal.toLong()
 
     fun toBaseMeasure(value: Double) = value * baseFactor
     fun fromBaseMeasure(value: Double) = value / baseFactor
